server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "google-auth>=2.49.1",
    "gunicorn>=25.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "PyMuPDF==1.27.1",
    "pymupdf-layout==1.27.1",
    "pymupdf4llm==0.3.4",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--strict-markers"

[tool.coverage.run]
source = [
    "authentication",
    "analytics",
    "core",
    "pdf_processing",
]
omit = [
    "*/migrations/*",
    "*/tests.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
